+++ cadastro-usuario/src/main/java/cadastrousuario/model/Usuario.java
@@ -14,7 +14,7 @@ import java.util.Date;
 public class Usuario {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long idUsuario;
+    private Integer idUsuario;
 
     private String nome;
